==> 1.Prepare/1.Prepare_transition_cb12.ipynb <==
import os, sys

import numpy as np
import pandas as pd

from collections import Counter, defaultdict

# Step 1: Load annotated job data
annoated by O*Net-SOC AutoCoder from user_history.tsv

df_title_annotated = pd.read_csv('../data/cb12/raw/title_annotated.csv', sep=';')

print(df_title_annotated.shape)
print("Unique JobTitle: ", len(df_title_annotated.JobTitle.unique()))
print("Unique JobTitle: ", len(df_title_annotated.JobTitle_token_new.unique()))
print("Unique Detailed Occupation: ", len(df_title_annotated.DetailedOccupation.unique()))

### Remove Code='99-9999'

df_title_annotated = df_title_annotated[df_title_annotated['DetailedOccupation'] != '99-9999'] 
print(df_title_annotated.shape)
print("Unique JobTitle: ", len(df_title_annotated.JobTitle.unique()))
print("Unique JobTitle: ", len(df_title_annotated.JobTitle_token_new.unique()))
print("Unique Detailed Occupation: ", len(df_title_annotated.DetailedOccupation.unique()))

### Remove Score < 70

df_title_annotated['Score_new'] = df_title_annotated['Score_new'].apply(lambda s: float(s))
df_title_annotated = df_title_annotated[df_title_annotated['Score_new'] >= 70] 
print(df_title_annotated.shape)
print("Unique JobTitle: ", len(df_title_annotated.JobTitle.unique()))
print("Unique JobTitle: ", len(df_title_annotated.JobTitle_token_new.unique()))
print("Unique Detailed Occupation: ", len(df_title_annotated.DetailedOccupation.unique()))

### Remove rare MinorGroup

dict_count_MinorGroup = df_title_annotated.MinorGroup.value_counts().to_dict()
MinorGroup_retain = [MinorGroup for MinorGroup, times in dict_count_MinorGroup.items() if times >= 200]
print(len(MinorGroup_retain))

df_title_annotated = df_title_annotated[df_title_annotated.MinorGroup.isin(MinorGroup_retain)]

df_title_annotated.shape

# Step 3: Prepare transition data

### Combine all infor

all_UserID = [] #list of UserID
all_JobTitle = [] #list of JobTitle
all_DetailedOccupation = [] #list of DetailedOccupation
all_MinorGroup = [] #list of MinorGroup
all_MajorGroup = [] #list of MajorGroup
all_items = []


for index, rows in df_title_annotated.groupby('UserID'):
    UserID = rows['UserID'].unique()[0]
    all_UserID.append(UserID)
    
    JobTitle_thisResume = [] #JobTitle items in a resume
    DetailedOccupation_thisResume = [] #DetailedOccupation items in a resume
    MinorGroup_thisResume = [] #MinorGroup items in a resume
    MajorGroup_thisResume = [] #MajorGroup items in a resume 
    
    for title, MajorGroup, MinorGroup, DetailedOccupation in zip(rows['JobTitle_token_new'], rows['MajorGroup'], rows['MinorGroup'], rows['DetailedOccupation']):
        JobTitle_thisResume.append(title)
        DetailedOccupation_thisResume.append(DetailedOccupation)
        MinorGroup_thisResume.append(MinorGroup)
        MajorGroup_thisResume.append(MajorGroup)
        
    
    all_JobTitle.append(JobTitle_thisResume)
    all_DetailedOccupation.append(DetailedOccupation_thisResume)
    all_MinorGroup.append(MinorGroup_thisResume)
    all_MajorGroup.append(MajorGroup_thisResume)

transition = {"UserID": all_UserID, 
              "JobTitle": all_JobTitle, 
              "DetailedOccupation": all_DetailedOccupation,
              "MinorGroup": all_MinorGroup,
              "MajorGroup": all_MajorGroup}
df_transition_raw = pd.DataFrame(transition)
print(df_transition_raw.shape)

def Combine_allInf(DetailedOccupations, MinorGroups, MajorGroups, JobTitles):
    combined_EmploymentItems = []
    for DetailedOccupation, MinorGroup, MajorGroup, JobTitle in zip(DetailedOccupations, MinorGroups, MajorGroups, JobTitles):
        combined_EmploymentItem = (DetailedOccupation, MinorGroup, MajorGroup, JobTitle)
        combined_EmploymentItems.append(combined_EmploymentItem)
    
    return combined_EmploymentItems
    
df_transition_raw['Combined_EmploymentItems'] = df_transition_raw.apply(lambda row: Combine_allInf(row['DetailedOccupation'], row['MinorGroup'], row['MajorGroup'], row['JobTitle']), axis = 1)

print(df_transition_raw.shape)

### Remove instances with Valid_EmploymentItems < 2

df_transition_raw['Num_EmploymentItems'] = df_transition_raw.Combined_EmploymentItems.apply(lambda x: len(x))
df_transition_final = df_transition_raw[df_transition_raw['Num_EmploymentItems'] > 1] 
print(df_transition_final.shape)

df_transition_final.to_csv('../data/cb12/processed/job_transition_MinorGroup200.csv', sep=';')

df_transition = pd.read_csv('./data/cb12/processed/job_transition_MinorGroup200.csv', sep=';')

df_transition
